// feature-1.14.0/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface CreatePlan {
    deliver_time: string;
    demand_ids: Array<number>;
    name: string;
    product_id: Record<string, any> | number | null;
    version: string;
    demand_docs: Array<{
      name: string;
      size: number;
      type: number;
      url: string | null;
      remark: string;
    }>;
    delay_list: Array<{ bug_id: number; iteration_id: number }>;
  }
  interface UpdatePlan {
    id: number | null;
    deliver_time: string;
    demand_ids: Array<number>;
    name: string;
    product_id: Record<string, any> | number | null;
    version: string;
    demand_docs: Array<{
      name: string;
      size: number;
      type: number;
      url: string | null;
      remark: string;
    }>;
    delay_list: Array<{ bug_id: number; iteration_id: number }>;
  }
// ...
}